Apply improved types & architecture to VeChain
