+++ libs/ledger-live-common/src/families/vechain/bridge/js.ts
@@ -1,14 +1,14 @@
-import type { Transaction } from "../types";
-import { makeAccountBridgeReceive } from "../../../bridge/jsHelpers";
-
-import { sync, scanAccounts } from "../js-synchronisation";
 import { AccountBridge, CurrencyBridge } from "@ledgerhq/types-live";
-import { createTransaction, prepareTransaction } from "../js-transaction";
-import getTransactionStatus from "../js-getTransactionStatus";
-import signOperation from "../js-signOperation";
-import broadcast from "../js-broadcast";
-import { estimateMaxSpendable } from "./js-estimateMaxSpendable";
 import { defaultUpdateTransaction } from "@ledgerhq/coin-framework/bridge/jsHelpers";
+import { makeAccountBridgeReceive } from "../../../bridge/jsHelpers";
+import { getTransactionStatus } from "../getTransactionStatus";
+import { estimateMaxSpendable } from "../estimateMaxSpendable";
+import { prepareTransaction } from "../prepareTransaction";
+import { createTransaction } from "../createTransaction";
+import { sync, scanAccounts } from "../synchronisation";
+import { signOperation } from "../signOperation";
+import { broadcast } from "../broadcast";
+import type { Transaction } from "../types";
 
 const receive: AccountBridge<Transaction>["receive"] = makeAccountBridgeReceive();
