+++ libs/ledger-live-common/src/families/vechain/synchronisation.ts
